progress on events dropdown list

+++ lib/Event.dart
@@ -2,8 +2,8 @@ class Event {
   final String eventName;
   final String date;
   final String time;
-  final double friends;
-  final double followers;
+  final int friends;
+  final int followers;
 
   Event(this.eventName, this.date, this.time, this.friends, this.followers);
 }
